requirements.txt
Django==4.1
djangorestframework==3.14.0
social-auth-app-django==5.4.0

